### 2018-07-24 #832 @arfon
:wave: @bmorris3. I'll be the handling editor for this submission.

Could you please take a look at this list of potential reviewers and see if there are any that might be suitable? https://bit.ly/joss-reviewers

Also, if you know of any potential reviewers _not_ on this list, feel free to mention them here!
