### 2018-07-24 #832 @bmorris3
👋 @arfon. Some potential reviewers include: @jakevdp, @manodeep, @wtgee
