### 2018-07-25 #832 @jakevdp
@manodeep or @wtgee, would you be willing to serve as a reviewer for this submission?
