### 2018-07-25 #832 @bmorris3
@manodeep has privately declined via Twitter, no response yet from @wtgee.
