### 2018-07-26 #832 @wtgee
Hi folks, thanks for asking. I can help out with the review but won't get
to it until this weekend.

On Thu, Jul 26, 2018 at 6:18 AM Néstor Espinoza <notifications@github.com>
wrote:

> Hi @bmorris3 <https://github.com/bmorris3>; would love to, but I'm
> swamped right now and currently already reviewing a paper. Might I
> reccommend @rabrahm <https://github.com/rabrahm> (
> https://github.com/rabrahm)?
>
