### 2018-07-26 #832 @manodeep
@bmorris3 @jakevdp My apologies for not being able to review. I know nothing about echelle spectra and only a little bit of python.
