### 2018-07-26 #832 @wtgee
I'll be following the [review process guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) but let me know if there is otherwise something I should know as this is my first JOSS review. Thanks!
