### 2018-07-26 301d793
Adding contributing file
