+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+# Contributing 
+
+[Pull requests](https://github.com/bmorris3/aesop/pulls), 
+[issues and comments](https://github.com/bmorris3/aesop/issues) are welcome by 
+all!  If you encouter a problem with the code, please contact us 
+by creating an [Issue via GitHub](https://github.com/bmorris3/aesop/issues).
+
+By participating in development of ``aesop``, we expect you to abide by the 
+[astropy Code of Conduct](http://www.astropy.org/code_of_conduct.html).
