### 2018-07-27 #854 @whedon
[REVIEW]: aesop: ARC Echelle Spectroscopic Observation Pipeline
**Submitting author:** @bmorris3 (<a href="http://orcid.org/0000-0003-2528-3409">Brett Morris</a>)
**Repository:** <a href="https://github.com/bmorris3/aesop" target ="_blank">https://github.com/bmorris3/aesop</a>
**Version:** v0.1
