+++ README.md
@@ -1,6 +1,6 @@
 # aesop
 
-[![Build Status](https://travis-ci.org/bmorris3/aesop.svg?branch=master)](https://travis-ci.org/bmorris3/aesop) [![Documentation Status](https://readthedocs.org/projects/arces/badge/?version=latest)](http://arces.readthedocs.io/en/latest/?badge=latest) [![Powered by Astropy Badge](http://img.shields.io/badge/powered%20by-AstroPy-orange.svg?style=flat)](http://www.astropy.org) 
+[![Build Status](https://travis-ci.org/bmorris3/aesop.svg?branch=master)](https://travis-ci.org/bmorris3/aesop) [![Documentation Status](https://readthedocs.org/projects/arces/badge/?version=latest)](http://arces.readthedocs.io/en/latest/?badge=latest) [![Powered by Astropy Badge](http://img.shields.io/badge/powered%20by-AstroPy-orange.svg?style=flat)](http://www.astropy.org) [![status](http://joss.theoj.org/papers/6737355abd6a7b0c20d22c7094576696/status.svg)](http://joss.theoj.org/papers/6737355abd6a7b0c20d22c7094576696)
 
 
 ARC Echelle Spectroscopic Observation Pipeline (aesop)
