### 2018-07-28 #854 @wtgee
@jakevdp I was recently also listed on the author's [astroplan paper](https://arxiv.org/abs/1712.09631). I'm looking at the conflict of interest policy and trying to decide if that applies but wanted to make you aware. Let me know if this presents an issue. Thanks!
