### 2018-07-29 #854 @bmorris3
I've read and re-read the COI terms and I think we're clear @wtgee.
