### 2018-08-06 #854 @wtgee
Hi @bmorris3 and @jakevdp. I've completed my review on this and everything looks good after the changes already made. @jakevdp I didn't hear an official response from you regarding the COI but have proceeded with the rest of the review. I do not perceive a conflict and neither does Brett.

Let me know what next steps I need to take as reviewer. Thanks!
