### 2018-08-09 #854 @jakevdp
@wtgee – let me double-check the COI
