### 2018-08-15 #854 @jakevdp
OK, the editorial board is in agreement that we're good to go here. Thanks everyone!
