### 2018-08-15 #854 @bmorris3
@arfon Done! http://doi.org/10.5281/zenodo.1345807
