### 2018-08-15 #854 @arfon
@wtgee - many thanks for your review here and thanks to @jakevdp for editing this submission ✨
